requirements.txt
streamlit==1.24.1
openai>=1.12.0
python-dotenv==1.0.0
pydantic==1.10.9
requests==2.31.0
